refactor: move 'isLocked' from BalanceSummary to fb.utils
